fix: absolute imports from the root dir work correctly both in localhost and build mode
In order for absolute imports in html or fetch files to work when in localhost server, the public dir must be set in vite config to be ../ (root resource), but when building, if the path is set like that, all aboslute imports in html that start with a / are automatically converted to be relative imports and get added a ".", but this fixes that. When hosting a localhost server, the public dir is set properly and all things load, but when compiling, the public dir is overwritten and disabled for the compiled code to work properly.